рефакторинг дан
