--- src/main/java/com/telebotZodiac/bot/entity/User.java
@@ -1,19 +1,0 @@
-//package com.telebotZodiac.bot.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.Data;
-
-//оставим если будем заводить юзера в базу данных (НЕ ТРОГАЕМ)
-//public class User {
-//    @Entity
-//    @Data
-//    public class Users {
-//        @Id
-//        @GeneratedValue
-//        private Long id;
-//        private String name;
-//        private Long chatId;
-//    }
-//}
